+++ ChangeLog.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `@compile_mode("unsupported")` for modules that do not support TorchScript
 
 ### Changed
+- `o3.TensorProduct` now uses `torch.fx` to generate it's code
+- e3nn now requires the latest stable PyTorch, >=1.8.0
 - in `soft_one_hot_linspace` the argument `base` is renamed into `basis`
 - `Irreps.slices()`, do `zip(irreps.slices(), irreps)` to retrieve the old behavior
 - `math.soft_one_hot_linspace` very small change in the normalization of `fourier` basis